refactor : presentation 디렉토리 분리
